refactor(InputTrigger): move out of KCD
I love JetBrains Rider.
