+++ osu.Game/Screens/Play/KeyCounter.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
@@ -79,21 +78,6 @@ namespace osu.Game.Screens.Play
             base.Dispose(isDisposing);
             Trigger.OnActivate -= Activate;
             Trigger.OnDeactivate -= Deactivate;
-        }
-
-        public abstract partial class InputTrigger : Component
-        {
-            public event Action<bool>? OnActivate;
-            public event Action<bool>? OnDeactivate;
-
-            protected InputTrigger(string name)
-            {
-                Name = name;
-            }
-
-            protected void Activate(bool forwardPlayback = true) => OnActivate?.Invoke(forwardPlayback);
-
-            protected void Deactivate(bool forwardPlayback = true) => OnDeactivate?.Invoke(forwardPlayback);
         }
     }
 }
